wallpepersphere/requirements.txt
blinker==1.7.0
boto3
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
google-genai
gunicorn
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Pillow
python-dotenv
SQLAlchemy==2.0.25
Werkzeug==3.0.1
google-auth
google-auth-oauthlib
requests
google-cloud-aiplatform
